+++ index.html
@@ -3,12 +3,14 @@
 
 <head>
     <title>optimaze-viewer-example</title>
-    <script src="build/bundle.js"></script>
     <link rel="stylesheet" href="build/bundle.css" />
 </head>
 
 <body>
+    <button id="login">Log in</button>
     <div id="viewer"></div>
+
+    <script src="build/bundle.js"></script>
 </body>
 
 </html>
